# framework/employee_mgmt/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "API documentation",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate a user",
                "description": "Logs in a user and creates a cookied-based session. Implements account lockout and status checks. Email is case insensitive and normalized to lowercase before authentication.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email address (case insensitive, normalized to lowercase)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "<set this in your .env file>"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "<set this in your .env file>"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials or unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Account is inactive or locked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Your account is inactive. Please contact support."
                                                },
                                                "error": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Too many failed login attempts. Your account is now locked. Please reset your password."
                                                },
                                                "error": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "expires_at": {
                                                    "type": "integer",
                                                    "example": 1701234567
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Your account is locked. Try again later or reset your password."
                                                },
                                                "error": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "expires_at": {
                                                    "type": "integer",
                                                    "example": 1701234567
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors occurred."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email field is required."
                                                ],
                                                "password": [
                                                    "The password field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred during login."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Retrieve a paginated list of companies",
                "description": "Fetches companies with pagination, search, and sorting. See CompanyFilterRequest and CompanySortRequest schemas for options.",
                "operationId": "getCompanies",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/itemsPerPageQueryParam"
                    },
                    {
                        "$ref": "#/components/parameters/pageQueryParam"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search by company name, EIN, or site names",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort: field direction. Fields: name, ein, status, employees_count, job_listings_count, job_applications_count, users_count, time_clocks_count, created_at",
                        "schema": {
                            "type": "string",
                            "example": "name asc"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Filter by company ID(s)",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved companies.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Companies retrieved successfully."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CompanyListResource"
                                                    }
                                                },
                                                "meta": {
                                                    "type": "object",
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/PaginationMeta"
                                                        },
                                                        {
                                                            "properties": {
                                                                "searchable_fields": {
                                                                    "description": "Fields that can be searched via the q parameter",
                                                                    "type": "array",
                                                                    "items": {
                                                                        "type": "string"
                                                                    },
                                                                    "example": [
                                                                        "name",
                                                                        "ein"
                                                                    ]
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to get companies."
                                        },
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Database connection failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompanyListResource": {
                "description": "Company list resource (list view)",
                "required": [
                    "id",
                    "company_type_id",
                    "name",
                    "ein",
                    "is_active"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9e7a662a-2cc7-43fb-8877-88a400d55788"
                    },
                    "company_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "company_type": {
                        "description": "Loaded relationship",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Corporation"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00+00:00"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T00:00:00+00:00"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Acme Corp"
                    },
                    "ein": {
                        "type": "string",
                        "example": "12-3456789"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "sites": {
                        "description": "Company sites",
                        "type": "array",
                        "items": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Main Office"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "employees_count": {
                        "type": "integer",
                        "example": 75,
                        "nullable": true
                    },
                    "job_listings_count": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "job_applications_count": {
                        "description": "All job applications for the company, including every status",
                        "type": "integer",
                        "example": 100,
                        "nullable": true
                    },
                    "candidates_count": {
                        "type": "integer",
                        "example": 25,
                        "nullable": true
                    },
                    "users_count": {
                        "description": "Active workspace users for the company; excludes Candidate/Employee-only and Rows Admin/Support users scoped here",
                        "type": "integer",
                        "example": 50,
                        "nullable": true
                    },
                    "time_clocks_count": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00+00:00",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00+00:00",
                        "nullable": true
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "LoginCompanyDepartmentSummary": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Transforms the authenticated user and their context into the login response payload.",
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Engineering Department"
                    }
                },
                "type": "object"
            },
...
            "LoginCompanySummary": {
                "required": [
                    "id",
                    "name",
                    "departments",
                    "sites"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Company Name"
                    },
                    "logo": {
                        "type": "string",
                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...",
                        "nullable": true
                    },
                    "departments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LoginCompanyDepartmentSummary"
                        }
                    },
                    "sites": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LoginCompanySiteSummary"
                        }
                    }
                },
                "type": "object"
            },
            "LoginUserContext": {
                "description": "Authenticated user with organization context and nested company summaries",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "email",
                    "companies"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "organization_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "organization_name": {
                        "type": "string",
                        "example": "RowsOne Inc."
                    },
                    "organization_logo": {
                        "type": "string",
                        "example": "https://example.com/logo.png",
                        "nullable": true
                    },
                    "companies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LoginCompanySummary"
                        }
                    }
                },
                "type": "object"
            },
            "LoginResource": {
                "title": "Login Bootstrap Payload",
                "description": "Authenticated user context returned by `/api/login` on the SDET challenge branch.",
                "required": [
                    "user"
                ],
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/LoginUserContext"
                    }
                },
                "type": "object"
            },
            "LoginSuccessResponse": {
                "title": "Login Success Response",
                "required": [
                    "message",
                    "error",
                    "data"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Login successful."
                    },
                    "error": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {
                        "$ref": "#/components/schemas/LoginResource"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "description": "Pagination metadata emitted under `data.meta` (Laravel paginator meta).",
                "properties": {
                    "current_page": {
                        "description": "Reusable OpenAPI schemas for paginated API responses.\n\nTwo response shapes exist:\n- `PaginationMeta` is emitted under `data.meta` by {@see ApiResponse::paginatedWithMeta()}.\n- `PaginationFlatMeta` is emitted as flat siblings of `data.items` by {@see ApiResponse::paginated()}.",
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "per_page": {
                        "description": "Items per page. `0` when fetching all items (itemsPerPage=*) with an empty result set.",
                        "type": "integer",
                        "example": 10
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 3
                    },
                    "total": {
                        "type": "integer",
                        "example": 25
                    },
                    "path": {
                        "type": "string",
                        "example": "https://api.example.com/api/resource"
                    },
                    "first_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "last_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string"
                                },
                                "active": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "pageQueryParam": {
                "name": "page",
                "in": "query",
                "description": "Page number (ignored when itemsPerPage is '*')",
                "required": false,
                "schema": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                }
            },
            "itemsPerPageQueryParam": {
                "name": "itemsPerPage",
                "in": "query",
                "description": "Number of items per page (use '*' to get all items, max 100)",
                "required": false,
                "schema": {
                    "type": "string",
                    "example": "10"
                }
            }
        }
    },
...
}